remove cruft and stuff

+++ source/player.c
@@ -1,13 +1,12 @@
-#include <stdint.h>
+#include "player.h"
 
 #include <gba.h>
+#include <stdint.h>
 
 #include "dangerous_dave.h"
 
-#include "player.h"
-
 #define WALK_SEQ_LEN 4
-const int_fast16_t walk_seq[WALK_SEQ_LEN] = {0, 1, 0, -1};
+const int walk_seq[WALK_SEQ_LEN] = {0, 1, 0, -1};
 
 // Sprite IDs + animations
 #define JUMPING_RIGHT 24
@@ -23,33 +22,35 @@ const int_fast16_t walk_seq[WALK_SEQ_LEN] = {0, 1, 0, -1};
 #define HITBOX_HEIGHT 16
 #define HITBOX_OFF_Y 0
 
-void player_loadSprites() {
+void player_loadSprites()
+{
     dmaCopy(dangerous_davePal, SPRITE_PALETTE, dangerous_davePalLen);
     dmaCopy(dangerous_daveTiles, SPRITE_GFX, dangerous_daveTilesLen);
 }
 
-void player_init(Player* p, OBJATTR* attribs) {
+void player_init(Player *p, OBJATTR *attribs)
+{
     p->spriteAttribs = attribs;
 
     /* set up the first attribute */
     p->spriteAttribs->attr0 = CENTER_PLAYER_SCREEN_Y | /* y coordinate */
-        (0 << 8) |                /* rendering mode */
-        (0 << 10) |               /* gfx mode */
-        (0 << 12) |               /* mosaic */
-        (1 << 13) |               /* color mode, 0:16, 1:256 */
-        (OBJ_SQUARE);             /* shape */
+                              (0 << 8) |               /* rendering mode */
+                              (0 << 10) |              /* gfx mode */
+                              (0 << 12) |              /* mosaic */
+                              (1 << 13) |              /* color mode, 0:16, 1:256 */
+                              (OBJ_SQUARE);            /* shape */
 
     /* set up the second attribute */
     p->spriteAttribs->attr1 = CENTER_PLAYER_SCREEN_X | /* x coordinate */
-        (0 << 9) |          /* affine flag */
-        (0 << 12) |         /* horizontal flip flag */
-        (0 << 13) |         /* vertical flip flag */
-        (ATTR1_SIZE_16);  /* size */
+                              (0 << 9) |               /* affine flag */
+                              (0 << 12) |              /* horizontal flip flag */
+                              (0 << 13) |              /* vertical flip flag */
+                              (ATTR1_SIZE_16);         /* size */
 
     /* setup the second attribute */
-    p->spriteAttribs->attr2 = 0 |   // tile index */
-        (0 << 10) | // priority */
-        (0 << 12);         // palette bank (only 16 color)*/
+    p->spriteAttribs->attr2 = 0 |         // tile index */
+                              (0 << 10) | // priority */
+                              (0 << 12);  // palette bank (only 16 color)*/
 
     p->firstFrame = 0;
     p->animFrame = 0;
@@ -71,38 +72,45 @@ void player_init(Player* p, OBJATTR* attribs) {
     p->hitbox.height = HITBOX_HEIGHT;
 }
 
-void player_destroy(Player* player) {
+void player_destroy(Player *player)
+{
     player->isJumping = false;
     player->firstFrame = 0;
     player->animFrame = 0;
     player->isMoving = false;
 }
 
-void player_setPosition(Player* this, int_fast16_t x, int_fast16_t y) {
+void player_setPosition(Player *this, int x, int y)
+{
     this->position.x = x;
     this->position.y = y;
     this->hitbox.x = this->position.x + HITBOX_OFF_X;
     this->hitbox.y = this->position.y + HITBOX_OFF_Y;
 }
 
-void player_tick(Player* player, int_fast16_t delay) {
+void player_tick(Player *player, int delay)
+{
     // only animate once every 8th frame
-    if ((delay % 16) > 0) {
+    if ((delay % 16) > 0)
+    {
         return;
     }
 
-    static int_fast16_t anim_seq;
+    static int anim_seq;
 
-    if (player->isMoving) {
+    if (player->isMoving)
+    {
         // if (player->isJumping) {
         //     player->firstFrame = player->facingLeft ? JUMPING_LEFT : JUMPING_RIGHT;
         //     player->animFrame = 0;
         // } else {
-            player->firstFrame = player->facingLeft ? WALKING_LEFT : WALKING_RIGHT;
-            anim_seq = ++anim_seq % WALK_SEQ_LEN;
-            player->animFrame = walk_seq[anim_seq];
+        player->firstFrame = player->facingLeft ? WALKING_LEFT : WALKING_RIGHT;
+        anim_seq = ++anim_seq % WALK_SEQ_LEN;
+        player->animFrame = walk_seq[anim_seq];
         // }
-    } else {
+    }
+    else
+    {
         player->firstFrame = player->facingLeft ? STANDING_LEFT : STANDING_RIGHT;
         player->animFrame = 0;
     }
@@ -110,7 +118,8 @@ void player_tick(Player* player, int_fast16_t delay) {
     player->spriteAttribs->attr2 = player->firstFrame + (player->animFrame * 8);
 }
 
-Vec player_getScreenPos(Player* this) {
+Vec player_getScreenPos(Player *this)
+{
     Vec vec;
     // Bitwise AND to get the lowest 8 bits
     vec.y = this->spriteAttribs->attr0 & 0x00FF;
@@ -119,10 +128,12 @@ Vec player_getScreenPos(Player* this) {
     return vec;
 }
 
-void player_setScreenX(Player* this, u16 x) {
+void player_setScreenX(Player *this, u16 x)
+{
     this->spriteAttribs->attr1 = (this->spriteAttribs->attr1 & 0xFE00) | (0x01FF & x);
 }
 
-void player_setScreenY(Player* this, u16 y) {
+void player_setScreenY(Player *this, u16 y)
+{
     this->spriteAttribs->attr0 = (this->spriteAttribs->attr0 & 0xFF00) | (0x00FF & y);
 }
