+++ source/player.h
@@ -12,9 +12,9 @@
 typedef struct Player {
     OBJATTR* spriteAttribs;
 
-    uint_fast16_t firstFrame;
-    uint_fast16_t animFrame;
-    uint_fast16_t framesInAir;
+    uint firstFrame;
+    uint animFrame;
+    uint framesInAir;
 
     bool isJumping;
     bool facingLeft;
@@ -31,22 +31,22 @@ void player_init(Player* p, OBJATTR* attribs);
 
 void player_destroy(Player* player);
 
-void player_tick(Player* player, int_fast16_t delay);
+void player_tick(Player* player, int delay);
 
-void player_setPosition(Player* this, int_fast16_t x, int_fast16_t y);
+void player_setPosition(Player* this, int x, int y);
 
 // Spritesheet information
 #define PLAYER_SPRITE_HEIGHT 16
 #define PLAYER_SPRITE_WIDTH 16
 
-// extern const u16 CENTER_PLAYER_SCREEN_X = (SCREEN_WIDTH / 2) - (PLAYER_SPRITE_WIDTH / 2);
-// extern const u16 CENTER_PLAYER_SCREEN_Y = SCREEN_HEIGHT - (PLAYER_SPRITE_HEIGHT * 3);
+// CENTER_PLAYER_SCREEN_X = (SCREEN_WIDTH / 2) - (PLAYER_SPRITE_WIDTH / 2);
 #define CENTER_PLAYER_SCREEN_X 112
+// CENTER_PLAYER_SCREEN_Y = SCREEN_HEIGHT - (PLAYER_SPRITE_HEIGHT * 3);
 #define CENTER_PLAYER_SCREEN_Y 112
 
-// extern const u16 MAX_PLAYER_SCREEN_X = SCREEN_WIDTH - PLAYER_SPRITE_WIDTH;
-// extern const u16 MAX_PLAYER_SCREEN_Y = SCREEN_HEIGHT - PLAYER_SPRITE_HEIGHT;
+// MAX_PLAYER_SCREEN_X = SCREEN_WIDTH - PLAYER_SPRITE_WIDTH;
 #define MAX_PLAYER_SCREEN_X 224
+// MAX_PLAYER_SCREEN_Y = SCREEN_HEIGHT - PLAYER_SPRITE_HEIGHT;
 #define MAX_PLAYER_SCREEN_Y 144
 
 Vec player_getScreenPos(Player* this);
